chore #17: 폴더 구조 변경 및 초기값 추가
