+++ PIMO/Sources/Model/Feed/TextImage.swift
@@ -12,4 +12,10 @@ struct TextImage: Decodable, Equatable, Hashable {
     let id: Int
     let imageURL: String
     let text: String
+    
+    init(id: Int = 0, imageURL: String = "", text: String = "") {
+        self.id = id
+        self.imageURL = imageURL
+        self.text = text
+    }
 }
